Contact Page baseline implemented
Baseline form for contact page is implemented, submitting a form that is validated in the backend will now publish the data to the newly implemented ContactTable.

All left to do is produce the admin view page. and to implement both of these pages with the html template provided by the front end team.